requirements.txt
# Основные для Python 3.12
aiogram==3.10.0
aiohttp==3.9.1
python-dotenv==1.0.0

# База данных
aiosqlite==0.20.0
sqlalchemy==2.0.23
greenlet==3.0.3

# Утилиты
loguru==0.7.2
qrcode[pil]==7.4.2
pillow==10.1.0
cryptography==41.0.7
python-dateutil==2.8.2
uuid==1.30

# Для работы с JSON
ujson==5.8.0

# Для асинхронных задач
asyncio-throttle==1.0.2

# Для SSH подключения к VPS
paramiko==3.4.0

# Для тестирования
pytest==7.4.3
pytest-asyncio==0.21.1
